# mongo/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NanoXplore RH — API Backend
  version: 1.0.0
  description: |
    API REST du backend NanoXplore RH (Express + MongoDB).
    Authentification par cookie JWT httpOnly (`token`).
    Base URL de montage : `/api` — le serveur écoute sur `http://localhost:3000` par défaut (PORT env).
servers:
  - url: /api
    description: Express API (chemin relatif au serveur)
security: []  # défaut public — chaque opération précise sa sécurité
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        JWT signé HS256, émis par POST /auth/login.
        Durée : 8h par défaut, 30j si remember=true.
        Cookie httpOnly, secure (production), sameSite=strict.

# =============================================================================
# SHARED PARAMETERS
# =============================================================================
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Numéro de page (base 1)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
      description: Nombre d'éléments par page (max 100)
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: ObjectId MongoDB (24 caractères hexadécimaux)

# =============================================================================
# SCHEMAS
# =============================================================================
  schemas:
    Error:
      type: object
      description: Format d'erreur standard — toutes les réponses 4xx/5xx.
      required: [error]
      properties:
        error:
          type: string
          description: Message d'erreur en français
      example:
        error: "Utilisateur introuvable"
    PaginatedResponse:
      type: object
      description: Enveloppe de réponse pour tous les endpoints de liste.
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items: {}
          description: Tableau d'objets de la page courante
        total:
          type: integer
          description: Nombre total d'éléments (toutes pages confondues)
        page:
          type: integer
          description: Page courante
        limit:
          type: integer
          description: Taille de page demandée
    User:
      type: object
      description: |
        Utilisateur de l'application. passwordHash et ldapDn ne sont jamais retournés.
        Le champ `id` (string) remplace `_id` dans la réponse de GET /auth/me.
      properties:
        _id:
          type: string
          description: ObjectId MongoDB
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, hr, director, manager, employee]
          description: |
            Rôles actifs : admin, hr, manager, employee.
            director : rôle legacy — traité comme manager dans l'architecture cible.
        department:
          type: string
          nullable: true
          enum:
            - Engineering
            - Product
            - Design
            - Data
            - Security
            - Infrastructure
            - Finance
            - Legal
            - HR
            - Sales
            - Marketing
            - "Customer Success"
            - Operations
            - Executive
        position:
          type: string
          nullable: true
          maxLength: 150
        managerId:
          type: string
          nullable: true
          description: ObjectId du manager direct (ref User)
        authSource:
          type: string
          enum: [local, ldap]
          description: "local = compte créé manuellement ; ldap = authentification via annuaire"
        locale:
          type: string
          enum: [fr, en]
        theme:
          type: string
          enum: [dark, light, light-sidebar]
        notificationPrefs:
          type: object
          description: |
            Clés booléennes filtrées par rôle.
            Clés possibles : campaignLaunch, evaluationAssigned, evaluationSubmitted,
            deadlineReminder, managerActionRequired, systemAlerts.
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        onboarding:
          type: object
          properties:
            completed:
              type: boolean
            completedAt:
              type: string
              format: date-time
              nullable: true
            steps:
              type: array
              items:
                type: object
                properties:
                  step:
                    type: string
                  done:
                    type: boolean
                  doneAt:
                    type: string
                    format: date-time
                    nullable: true
        offboardingStatus:
          type: string
          enum: [active, offboarding, offboarded]
        offboardingDate:
          type: string
          format: date-time
          nullable: true
        offboardingReason:
          type: string
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OffboardingRequest:
      type: object
      description: |
        Demande de départ (unique par utilisateur — index unique sur userId).
        La completion (status=completed) archive l'utilisateur (isActive=false).
      properties:
        _id:
          type: string
        userId:
          description: Utilisateur qui part (peuplé dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        requestedBy:
          description: Auteur de la demande (peuplé dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        reason:
          type: string
          enum: [resignation, termination, retirement, other]
        lastDay:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, in_progress, completed]
        checklist:
          type: array
          description: Liste de vérification par défaut (5 items)
          items:
            type: object
            properties:
              _id:
                type: string
              item:
                type: string
              done:
                type: boolean
              doneAt:
                type: string
                format: date-time
                nullable: true
              doneBy:
                type: string
                nullable: true
        notes:
          type: string
          nullable: true
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /users/{id}/offboard:
    patch:
      summary: Déclencher un départ
      description: Marque l'utilisateur en offboarding et archive ses évaluations non terminales (admin/hr).
      tags: [users]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, effectiveDate]
              properties:
                reason:
                  type: string
                  description: Motif du départ
                effectiveDate:
                  type: string
                  format: date
                  description: Date effective du départ (ISO 8601)
      responses:
        '200':
          description: Offboarding déclenché — utilisateur retourné avec offboardingStatus=offboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Champs requis manquants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux admins et RH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /offboarding:
    get:
      summary: Liste des demandes de départ
      description: Retourne les demandes paginées, filtrables par statut (hr/admin).
      tags: [offboarding]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
      responses:
        '200':
          description: Liste paginée de demandes de départ
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OffboardingRequest'
        '403':
          description: Réservé aux hr et admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Créer une demande de départ
      description: Crée une demande de départ (hr/admin). Unique par utilisateur (index unique sur userId).
      tags: [offboarding]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, reason, lastDay]
              properties:
                userId:
                  type: string
                  description: ObjectId de l'utilisateur qui part
                reason:
                  type: string
                  enum: [resignation, termination, retirement, other]
                lastDay:
                  type: string
                  format: date
                notes:
                  type: string
                  maxLength: 2000
      responses:
        '201':
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffboardingRequest'
        '400':
          description: Champs requis manquants, reason invalide ou userId non ObjectId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux hr et admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Demande déjà existante pour cet utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /offboarding/{id}:
    get:
      summary: Détail d'une demande de départ
      description: Retourne une demande de départ par son ID (hr/admin).
      tags: [offboarding]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Demande trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffboardingRequest'
        '403':
          description: Réservé aux hr et admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Mettre à jour une demande de départ
      description: |
        Modifie le statut ou les notes (hr/admin).
        status=completed → user.isActive=false, user.offboardingStatus=offboarded.
      tags: [offboarding]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, in_progress, completed]
                notes:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: Demande modifiée (peuplée avec userId et requestedBy)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffboardingRequest'
        '400':
          description: status invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux hr et admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Supprimer une demande de départ
      description: Supprime une demande de départ (admin uniquement).
      tags: [offboarding]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Demande supprimée
        '403':
          description: Réservé à l'administrateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /offboarding/{id}/checklist/{itemIndex}:
    patch:
      summary: Cocher un item de la checklist
      description: |
        Marque un item de la checklist comme fait (done=true) ou non (hr/admin).
        Premier item coché → statut passe automatiquement pending→in_progress.
      tags: [offboarding]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: itemIndex
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
          description: Index (0-based) de l'item dans la checklist
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                done:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Checklist mise à jour (demande peuplée retournée)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OffboardingRequest'
        '400':
          description: itemIndex invalide ou hors limites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux hr et admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------------------------------------------------------------------------
  # Admin
  # ---------------------------------------------------------------------------
...